Add community guide for lazy servers
This guide details how to set up and use a lazy server using `lazymc`
while providing access to the BlueMap web interface even when the server
is offline.

It uses Nginx to host the BlueMap web interface in the same way as the
existing wiki article details, while extending it to also return fake
"live" data to keep the frontend happy when the server is offline.