Update lazy server guide to handle lazymc's ability to pause servers (#21)
Lazymc can now use `SIGSTOP` and `SIGCONT` to pause and resume servers
instead of just shutting them down.

While in a paused state the integrated webserver will accept incoming
requests, but not be able to respond to them. This results in requests
that hang for a long time (60s by default) before returning a 504. This
commit updates the guide to handle this case by recommending to set a
low `proxy_read_timeout` if the pause method is used.