Update Gemfile.lock
first time ruby user and project contributor here, not really sure if this is "ok" or not?
if you want this commit removed from the PR im more than happy to do so.
happened after running `bundle update`, which it seemed like I needed to do to get the webserver running, because of a missing dependency that bundle could not locate (missing/removed)

err msg:
```
Your bundle is locked to ffi (1.15.5-x64-mingw-ucrt) from rubygems repository https://rubygems.org/ or installed locally, but that version can no longer be found in that source
```