docs(wiki): fix typos, improve formatting/readability (#129)
* docs(wiki): add lazymc readme link

* docs(wiki): improve readability, formatting, typos

- fixed some of the formatting around the wiki (a few lists weren't displaying correctly, long paragraphs that look like they were intended to be split over multiple lines, etc)
- fixed several typos, mainly with the word "dependencies". Plus some grammar/punctuation while I was in there
- removed quite a few unnecessary whitespaces that do not effect the page layout/formatting
- fixed up some links that were either not linking, or had poor use of "[example.tld](https://example.tld)" (like quoting the whole target URL inside the link text instead of using `<https://example.tld>`?)

* Update Gemfile.lock

first time ruby user and project contributor here, not really sure if this is "ok" or not?
if you want this commit removed from the PR im more than happy to do so.
happened after running `bundle update`, which it seemed like I needed to do to get the webserver running, because of a missing dependency that bundle could not locate (missing/removed)

err msg:
```
Your bundle is locked to ffi (1.15.5-x64-mingw-ucrt) from rubygems repository https://rubygems.org/ or installed locally, but that version can no longer be found in that source
```

* Update community/Chat.md

---------

Co-authored-by: Lukas Rieger <TBlueF@users.noreply.github.com>